Make adding sections feel smoother
Use a button to add a section with an empty header name to the start of the list, and then focus on the first item so users can add a text immediately. Header input is shown when there are more than 1 task in a section. Automatically makes the header the first task if there are no items in the list. Click away to save changes.